requirements.txt
selenium==4.18.1
python-dotenv==1.0.1
schedule==1.2.1
python-telegram-bot==20.8
webdriver-manager==4.0.1
requests>=2.26.0
pytest==8.0.2
black==24.2.0
flake8==7.0.0
flask==2.3.3
flask-sqlalchemy==3.1.1
flask-wtf==1.2.1
flask-login==0.6.3
email-validator==2.1.0
icalendar==6.1.3
psycopg2-binary==2.9.9
alembic==1.13.1 